requirements.txt
streamlit
llama-index
llama-index-llms-llama-api
python-dotenv
